Remove per-class PropertyDefaults in favour of a single object
