+++ test-schemas-domain-classes/src/main/scala/testdomains/codepropertygraphminified/PropertyDefaults.scala
@@ -1,0 +1,9 @@
+package testdomains.codepropertygraphminified
+
+object PropertyDefaults {
+  val DispatchType = "<empty>"
+
+  val Name = "<empty>"
+
+  val Order = -1: Int
+}
