+++ test-schemas-domain-classes/src/main/scala/testdomains/codepropertygraphminified/nodes/Method.scala
@@ -27,11 +27,11 @@ object Method {
   object Properties {
     val Name = flatgraph.SinglePropertyKey[String](kind = 1, name = "NAME", default = "<empty>")
   }
-  object PropertyDefaults {
-    val Name = "<empty>"
-  }
 }
 
+/** Node properties:
+  *   - Name
+  */
 class Method(graph_4762: flatgraph.Graph, seq_4762: Int)
     extends StoredNode(graph_4762, 1, seq_4762)
     with MethodBase
