+++ test-schemas-domain-classes/src/main/scala/testdomains/empty/PropertyDefaults.scala
@@ -1,0 +1,3 @@
+package testdomains.empty
+
+object PropertyDefaults {}
