+++ test-schemas-domain-classes/src/main/scala/testdomains/generic/PropertyDefaults.scala
@@ -1,0 +1,7 @@
+package testdomains.generic
+
+object PropertyDefaults {
+  val IntMandatory = 42: Int
+
+  val StringMandatory = "<empty>"
+}
