+++ test-schemas-domain-classes/src/main/scala/testdomains/gratefuldead/PropertyDefaults.scala
@@ -1,0 +1,7 @@
+package testdomains.gratefuldead
+
+object PropertyDefaults {
+  val Name = ""
+
+  val Weight = 0: Int
+}
