+++ test-schemas-domain-classes/src/main/scala/testdomains/gratefuldead/nodes/Artist.scala
@@ -27,11 +27,11 @@ object Artist {
   object Properties {
     val Name = flatgraph.SinglePropertyKey[String](kind = 0, name = "name", default = "")
   }
-  object PropertyDefaults {
-    val Name = ""
-  }
 }
 
+/** Node properties:
+  *   - Name
+  */
 class Artist(graph_4762: flatgraph.Graph, seq_4762: Int)
     extends StoredNode(graph_4762, 0, seq_4762)
     with ArtistBase
