+++ test-schemas-domain-classes/src/main/scala/testdomains/hierarchical/PropertyDefaults.scala
@@ -1,0 +1,5 @@
+package testdomains.hierarchical
+
+object PropertyDefaults {
+  val Name = "<empty>"
+}
